abstract sftpclient, progress monitor code
